code review feedback

+++ config/config.go
@@ -71,7 +71,7 @@ type Config struct {
 	HTTPReadTimeout             time.Duration             `yaml:"http_read_timeout"`
 	HTTPWriteTimeout            time.Duration             `yaml:"http_write_timeout"`
 	AccessLogLevel              string                    `yaml:"access_log_level"`
-	MaxBlobSize					int	                      `yaml:"max_blob_size"`
+	MaxBlobSize                 int64                     `yaml:"max_blob_size"`
 
 	// Fields that are created by combinations of the flags above.
 	ProxyBackend cache.Proxy
@@ -106,7 +106,7 @@ func newFromArgs(dir string, maxSize int, storageMode string,
 	httpReadTimeout time.Duration,
 	httpWriteTimeout time.Duration,
 	accessLogLevel string,
-	maxBlobSize int) (*Config, error) {
+	maxBlobSize int64) (*Config, error) {
 
 	c := Config{
 		Host:                        host,
@@ -384,6 +384,6 @@ func get(ctx *cli.Context) (*Config, error) {
 		ctx.Duration("http_read_timeout"),
 		ctx.Duration("http_write_timeout"),
 		ctx.String("access_log_level"),
-		ctx.Int("max_blob_size"),
+		ctx.Int64("max_blob_size"),
 	)
 }
